backend/requirements.txt
numpy
pandas
scipy
torchvision
torch
ultralytics
flask
flask-sqlalchemy
flask-cors
waitress
gdrive
pillow